Add support for the new aprsd-gps-extension
This patch adds support for the new aprsd-gps-extension.
The extension calls gpsd to fetch real GPS coordinates.

The new webchat UI has a new GPS button that expands an area
to set the GPS beaconing type.
0 = disable GPS Beaconing
1 = Manual beaconing (just like the old webchat)

- Automatic beaconing
2 = interval beaconing - beacon ever N seconds
3 = smart beaconing - beacon when you move X feet within Y timeframe

Currently the only working option is 1.   Still working on changing
the settings and making aprsd webchat and the gps extension do the
right thing.