Added automatic updates to location string
when a user has fetched the location of the callsign for a tab,
we now update the timeago string in the location_string.  So
over the course of time they will see that they got the location
X time ago.  Previously, this was statically built on the webchat.py
and would stay the same string forever.